Split files into two files. Rename cameraInterface.
